1st submittion of 179_lagestnumber
1st submittion of 179_lagestnumber, very low performance

+++ Source/LeeCode/Alogrithm/274_h-index.h
@@ -21,12 +21,13 @@ namespace Anlogrithm_274
         int hIndex(vector<int>& citations) {
 
             sort(citations.begin(), citations.end(), [] (int x, int y){return x > y; });
-            for (int i = 0; i < citations.size(); i++)
+            for (int i = 0; i < (int)citations.size(); i++)
             {   
                 if (i >= citations[i])
                     return i;
             }
             return citations.size();
+
         }
     };
